backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
feedparser==6.0.11
anthropic>=0.40.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.9.0
httpx
